# docs/api/voidengine/swagger.yaml
basePath: 这里写base path
consumes:
- application/json
definitions:
  message.AddUserRequest:
    properties:
      email:
        maxLength: 100
        minLength: 5
        type: string
      password:
        description: 密码
        maxLength: 20
        minLength: 8
        type: string
      phone:
        maxLength: 20
        minLength: 4
        type: string
      rePassword:
        maxLength: 20
        minLength: 8
        type: string
      username:
        description: 用户名
        maxLength: 32
        type: string
    required:
    - password
    - rePassword
    - username
    type: object
  message.DeleteUserRequest:
    properties:
      username:
        description: 用户名
        maxLength: 32
        type: string
    required:
    - username
    type: object
  message.ModifyUserRequest:
    properties:
      email:
        maxLength: 100
        minLength: 5
        type: string
      password:
        description: 密码
        maxLength: 20
        minLength: 8
        type: string
      phone:
        maxLength: 20
        minLength: 4
        type: string
      rePassword:
        maxLength: 20
        minLength: 8
        type: string
      username:
        description: 用户名
        maxLength: 32
        type: string
    required:
    - username
    type: object
  response.ApiResponse:
    properties:
      code:
        type: integer
      data: {}
      details: {}
      message:
        type: string
    type: object
host: 这里写接口服务的host
info:
  contact:
    email: support@swagger.io
    name: 这里写联系人信息
    url: http://www.swagger.io/support
  description: 这里写描述信息
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: '"API 文档"'
paths:
  /ping:
    get:
      consumes:
      - application/json
      description: Ping
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
      tags:
      - ping
    post:
      consumes:
      - application/json
      description: Ping
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
      tags:
      - ping
  /v1/traffic:
    get:
      consumes:
      - application/json
      description: List NetworkTraffics
      parameters:
      - default: zh
        description: Language
        in: header
        name: Accept-Language
        type: string
      - description: 排序方式[desc, asc]
        example: desc
        in: query
        maxLength: 4
        name: orderBy
        type: string
      - description: 当前页码
        example: 1
        in: query
        minimum: 1
        name: pageNo
        type: integer
      - description: 每页数量
        example: 50
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      tags:
      - NetworkTraffic
  /v1/user:
    delete:
      consumes:
      - application/json
      description: Delete a User
      parameters:
      - default: zh
        description: Language
        in: header
        name: Accept-Language
        type: string
      - description: delete user param
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/message.DeleteUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      tags:
      - User
    get:
      consumes:
      - application/json
      description: List Users
      parameters:
      - default: zh
        description: Language
        in: header
        name: Accept-Language
        type: string
      - description: 排序方式[desc, asc]
        example: desc
        in: query
        maxLength: 4
        name: orderBy
        type: string
      - description: 当前页码
        example: 1
        in: query
        minimum: 1
        name: pageNo
        type: integer
      - description: 每页数量
        example: 50
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      - description: 用户名
        in: query
        maxLength: 32
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Add a User
      parameters:
      - default: zh
        description: Language
        in: header
        name: Accept-Language
        type: string
      - description: add user param
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/message.AddUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Modify a User
      parameters:
      - default: zh
        description: Language
        in: header
        name: Accept-Language
        type: string
      - description: modify user param
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/message.ModifyUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      tags:
      - User
produces:
- application/json
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"